move/divide into submodules: languoids.* and references.*
