+++ pyglottolog/__init__.py
@@ -1,2 +1,7 @@
-#
+# pyglotolog
+
+from .api import Glottolog
+
+__all__ = ['Glottolog']
+
 __version__ = '1.0.0'
