+++ pyglottolog/languoids/__init__.py
@@ -1,0 +1,19 @@
+# languoids
+
+from __future__ import unicode_literals
+
+from .languoid import Languoid
+from .models import (
+    Glottocode, Glottocodes,
+    Level, Macroarea, Country, Reference,
+    ClassificationComment, EndangermentStatus,
+    EthnologueComment, ISORetirement,
+)
+
+__all__ = [
+    'Languoid',
+    'Glottocode', 'Glottocodes',
+    'Level', 'Macroarea', 'Country', 'Reference',
+    'ClassificationComment', 'EndangermentStatus',
+    'EthnologueComment', 'ISORetirement',
+]
