--- pyglottolog/monsterlib/__init__.py
@@ -1,2 +1,0 @@
-# coding: utf8
-from __future__ import unicode_literals, print_function, division
