+++ pyglottolog/references/__init__.py
@@ -1,0 +1,13 @@
+# references
+
+from __future__ import unicode_literals
+
+from .bibfiles import BibFiles, BibFile, Entry
+from .hhtypes import HHTypes
+from .isbns import Isbns, Isbn
+
+__all__ = [
+    'BibFiles', 'BibFile', 'Entry',
+    'Isbns', 'Isbn',
+    'HHTypes',
+]
