+++ pyglottolog/references/bibtex_undiacritic.py
@@ -1,10 +1,10 @@
-# _bibtex_undiacritic.py - remove latex/unicode diacritics for string comparison
+# bibtex_undiacritic.py - remove latex/unicode diacritics for string comparison
 
 import re
 
-from unidecode import unidecode
 from six import text_type
 
+from unidecode import unidecode
 
 COMMAND1 = re.compile(r'\\[a-z]+\{([^}]*)\}')
 COMMAND2 = re.compile(r'\\text[a-z]+')
