+++ pyglottolog/tests/util.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals, print_function, division
 from clldutils.path import Path, copytree
 from clldutils.testing import WithTempDir
 
-from pyglottolog.api import Glottolog
+from pyglottolog import Glottolog
 
 
 class WithApi(WithTempDir):
